refactor(libs-frontend): #1645 make FlexBoxEditor and its util modules
- fix: add guiCss to the check when creating the css in atomRenderPipe
