+++ libs/frontend/modules/element/src/css-editor/css-layout-editor/index.ts
@@ -1,0 +1,1 @@
+export * from './FlexBoxEditor'
