+++ libs/frontend/modules/element/src/css-editor/utils/index.ts
@@ -1,0 +1,2 @@
+export * from './makeMenu'
+export * from './updateGuiCssProperty'
